fix an issue with supertest
Workaround for: `TypeError: Property 'request' of object #<Context> is
not a function` described in a comment in issue #2

In supertest, request is no longer a function, but an object with
methods as its keys.

This check whether `this.request` is a Function. If so, do the usual,
otherwise assume it's an object.